Refactor Item entity
I have refactored the item entity, specially the way it's creation prerequisites are checked.
There are also minor changes on other entities.
More changes should be made to the verbs to fit the updated item. For now, there are uncatched
exceptions and poor error messages when trying to save wrong items.