give max priority to custom verbs - resolves #63
This way, a custom verb with the same name of a regular verb will be executed instead of the regular one.
This gives creators freedom to override the regular behavior of the world. In the future we might want to
put some "security" verbs higher in the priority list i.e. the verb used to exit a world.