update Item.get_items_in_world() - resolves #81
Now it takes into account items being carried at users' inventories.
This solution will need to be updated when we allow for multiple worlds as in #15
It also may be a source of overhead if we have a lot of users at once, since it
searchs for items in each user's inventory.