add item check command - resolves #77
add an "if" command that allows to perform different actions based
on the pressence of an item. Useful for custom verbs.

It is decided to force users to start the commands passed to the
"if" command with an "-". Since an "OK" line is used to finish every
block of actions, the preceding "-"s lets the algorithm work with
nested blocks of actions.