change saved items to belong to a world state.
Until now, saved items were global to the server. With this change,
saved items belong to the world state where they are saved, and
also are snapshoted and deployed with them. Users see the items
saved by all other users when using the Place verb.
This change is made in order to facilitate the implementation of
the export/import verbs. This way saved items will be exported
with the world, which is now self-contained.