enhance user target inference - resolves #108
Previously, each verb did the work of transforming an item or exit
name into the desired entity. This commits introduces the util
function name_to_entity, that implements the functionality of most
use cases.

This function has been used in most verb to handle the user target
inference, and now users can enter an incomplete name, case and
accent insensitive in most verbs that require selecting a entity
by name.

There are still verbs that don't use the function, and others that
need to enter the exact value (like delete verbs).

Further work is required to make this aspect perfect, but is enough
for now.