add password support - resolves #14
the sha256 hash function is applied to passwords to avoid storing
them as plain text. However, this approach is still insecure.
* Because salt is not applied, rainbow tables could be used to
retrieve the paswords.
* Because sha256 is a fast function, brute force attacks are easy.

In the future we should use one of the slower functions (> 100ms
per password). I didn't implement that right away because I'm
afraid that a slow function would make Architext to vulnerable to
DDoS attacks as it is now. Once there is a limit of password
attempts, we should upgrade the password hashing.