server: add queries proof of concept
I want to have the power to add raw SQL queries in the future.
For the first query I've created it as a protocol that is
implemented by a handler that uses the unit of work as data
source. This way I'll be able to substitute it by a SQL query
in production, but still test without a database.

To instantiate all future queries and pass them the uow I've added
a Query factory.