core: enforce the use of transactions
I have forgotten to use the uow within a with block
a lot of time. There are other cases where I forgot to
commit().

I hace created a facade around uow named Transaction,
that forces to use the uow within a with block. Also, commit
is not automatically called if the with block exits without
errors.